update react assignment 1 solution
