--- todo-list/src/App.js
@@ -1,13 +1,0 @@
-
-import './App.css';
-import Task from './Task';
-
-function App() {
-  return (
-    <>
-      <Task />
-    </>
-  );
-}
-
-export default App;
